General comments
Awesome project! I think this app is such a great idea. The design/style looks really nice. I appreciate that the website has a nice interface, it's simple to use, and loads quickly. My suggestions would be to just tweak a couple of things slightly, which I commented about in a few of the other files. Also, a dropdown/side navbar could be nice, since the top nav buttons disappear if the page is resized to be small. 
Great work!!